CommunicationError to Object with static methods
